Refine LogDeduplicator eviction coverage
